utils/gapy/requirements.txt
pyelftools>=0.24
argparse>=1.4
argcomplete
six

